fix(runtime): support TypeScript moduleResolution NodeNext
