+++ apps/router-demo/router-remote2-2002/src/bootstrap.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
-import App from './App';
+import App from './App.jsx';
 
 //@ts-ignore
 const root = createRoot(document.getElementById('root')!);
