+++ apps/router-demo/router-remote2-2002/src/export-App.tsx
@@ -1,4 +1,4 @@
-import App from './App';
+import App from './App.jsx';
 import { createBridgeComponent } from '@module-federation/bridge-react/v18';
 import { createInstance, getInstance } from '@module-federation/runtime';
